make public TReadoutBoard::GetReceiver() and minor modifications to reduce the warnings at compilation

+++ TScan.cpp
@@ -59,8 +59,8 @@ void TScan::CreateScanHisto ()
 
   THisto histo = CreateHisto ();
 
-  for (int iboard = 0; iboard < m_boards.size(); iboard ++) {
-    for (int ichip = 0; ichip < m_chips.size(); ichip ++) {
+  for (int iboard = 0; iboard < (int)m_boards.size(); iboard ++) {
+    for (int ichip = 0; ichip < (int)m_chips.size(); ichip ++) {
       if ((m_chips.at(ichip)->GetConfig()->IsEnabled()) && (m_chips.at(ichip)->GetReadoutBoard() == m_boards.at(iboard))) {
         id.boardIndex       = iboard;
         id.dataReceiver     = m_chips.at(ichip)->GetConfig()->GetParamValue("RECEIVER");
